kemampuan-dasar/bulan01/minggu05/cronos accounting & finance/requirements.txt
flask
python-dotenv
psycopg2_primary